+++ wutengine/src/context/mod.rs
@@ -3,6 +3,7 @@
 mod engine;
 mod gameobject;
 mod graphics;
+mod message;
 mod plugin;
 mod viewport;
 mod window;
@@ -10,6 +11,7 @@ mod window;
 pub use engine::*;
 pub use gameobject::*;
 pub use graphics::*;
+pub use message::*;
 pub use plugin::*;
 pub use viewport::*;
 pub use window::*;